% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avocado-package.R
\docType{package}
\name{avocado-package}
\alias{avocado}
\alias{avocado-package}
\title{Avocado Prices}
\format{Some relevant columns in the dataset:
\enumerate{
     \item \code{Data} - The date of the observation
     \item \code{AveragePrice} - the average price of a single avocado
     \item \code{type} - conventional or organic
     \item \code{year} - the year
     \item \code{Region} - the city or region of the observation
     \item \code{Total Volume} - Total number of avocados sold
     \item \code{4046} - Total number of avocados with PLU 4046 sold
     \item \code{4225} - Total number of avocados with PLU 4225 sold
     \item \code{4770} - Total number of avocados with PLU 4770 sold
}}
\usage{
avocado
}
\description{
The dataset contains historical data on avocado prices and
sales volume in multiple US markets.
\href{https://www.kaggle.com/neuromusic/avocado-prices/home}{Source Data} is from
\href{https://www.kaggle.com/}{kaggle}, for more relevant data, please check \href{http://www.hassavocadoboard.com/retail/volume-and-price-data}{here}
}
\examples{
head(avocado)
summary(avocado)
}
\keyword{internal}
