% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.reg.analysis.R
\name{dist.reg.map}
\alias{dist.reg.map}
\title{Avocado algorithm - Wall-to-wall map version}
\usage{
dist.reg.map(
  s,
  dates,
  rfd,
  dstrb_thr,
  rgrow_thr,
  cdates,
  nCluster,
  outname,
  format,
  datatype
)
}
\arguments{
\item{dates}{The julian dates for each scene in the time series brick. Should be in the following format: "YYYY-MM-DD"}

\item{rfd}{The reference frequency distribution. Determines if an anomaly falls outside the 95 percent of the reference frequency distribution. For example a value that fall in a RDF >= 0.95, indicates that the detected anomaly belongs to the 5 percent of lowest values and is a potential disturbance/change.}

\item{dstrb_thr}{sets a threshold (number of days) in which if a regrowth is detected within n days after a potential disturbance, then the candidate disturbance date is neglected.}

\item{rgrow_thr}{sets a threshold (number of days) in which if a disturbance is detected within n days after a potential regrowth, then the candidate regrowth date is neglected.}

\item{cdates}{Sets the number of consecutive data points for a change to be detected (minimum 2 and maximum 5 data points).}

\item{nCluster}{Numeric. Number of CPU's to be used for the job.}

\item{outname}{Character vector with the output path and filename with extension or only the filename and extension if work directory was set. More information: See writeRaster}

\item{format}{Character. Output file type. More information: See \code{\link{writeFormats}}}

\item{datatype}{Character vector that determines the interpretation of values written to disk. More information: See \code{\link{dataType}}}
}
\value{
The output will consist of a RasterStack with 16 bands: Band1=disturbance detection; Band2=corresponding anomalies band 1; Band3=regrowth detection; Band4=corresponding anomalies band 3; Band5= 2nd disturbance detection for data points that had a previous disturbance and regrowth; Band 6 etc.
}
\description{
Continuous vegetation change detection
}
\examples{
source("dist.reg.map.R") # Load in the mapping function
dates <- lan.dates # The dates from your time-series brick (x)
ano.rfd.st <- brick("ano.rfd.st.tif") #Load in the anomaly-rfd brick that you created in section 2
dist.reg.map(s=ano.rfd.st, dates=lan.dates, rfd=0.95, dstrb_thr=1, rgrow_thr=730,
nCluster=1,cdates=3,outname="YourDirectory/ChangeMap.tif", format="GTiff", datatype="INT2S")
}
\seealso{
\code{\link{dist.reg}}
}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
