% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir.R
\name{mkdir}
\alias{mkdir}
\title{Detect and create the directory of the output file.}
\usage{
mkdir(output)
}
\arguments{
\item{output}{A file's name.}
}
\value{
None
}
\description{
Detect and create the directory of the output file.
}
\examples{

# directory "foo/bar/" does not exist.
mkdir(output = "foo/bar/foobar")


}
