% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{n_levels}
\alias{n_levels}
\title{Get the number of unique levels of a vector.}
\usage{
n_levels(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{If TRUE, ignore NA.}
}
\value{
A number of levels.
}
\description{
Get the number of unique levels of a vector.
}
\examples{

n_levels(c(rep("A", 3), rep("B", 2), rep("C", 1)))

}
