% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLUGPhenAnoRFDPLUS.R
\name{PhenRef2d}
\alias{PhenRef2d}
\title{Avocado algorithm - calculate likelihood of vegetation-index–time space for reference vegetation}
\usage{
PhenRef2d(phen, dates, h, anop, rge)
}
\arguments{
\item{phen}{Numeric vector with the values of the reference vegetation}

\item{dates}{A date vector. The number of dates must be equal to the number of values of time series.}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of values that are in x. For those values the anomalies and likelihoods will be calculated based on the phen. For example a time series has 450 values, so anop=c(1:450).}

\item{rge}{A vector containing minimum and maximum values of the response variable used in the analysis. We suggest the use of theoretically based limits. For example in the case of MODIS NDVI or EVI, it ranges from 0 to 10,000, so rge =c(0,10000)}
}
\value{
List which contains cumulative bivariate density distribution (cumDensity), and maximum likelihood of the vegetation-index–time space (MAXY)
}
\description{
Calculate the cumulative density distribution and their likelihood values based on reference vegetation - input for Wall-to-wall map version (i.e. PLUGPhenAnoRFDPLUS function)
}
\examples{
\dontrun{
# Loading raster data
library(terra)
library(npphen)

MDD <- rast(system.file("extdata", "MDD_NDMI_1990_2020.grd", package = "AVOCADO"))
# load dates vector
load(system.file("extdata", "MDD_dates.RData", package = "AVOCADO"))
# load  reference forest shapefile
load(system.file("extdata", "MDDref.RData", package = "AVOCADO"))
# Create the reference curve
ref.ext <- ext(MDDref)
ref.brick <- crop(MDD, ref.ext)
fin <- nrow(ref.brick) * ncol(ref.brick)
phen <- as.numeric(terra::extract(ref.brick, 1))
d1 <- MDD_dates
for (i in 2:fin) {
  pp <- as.numeric(terra::extract(ref.brick, i))
  phen <- c(phen, pp)
  d1 <- c(d1, MDD_dates)
}
MDD_ref <- PhenRef2d(phen, d1, h = 1, anop = c(1:1063), rge = c(0, 10000))

# plot reference curve + probabilities (same result as using PhenKplot())
image(MDD_ref$x, MDD_ref$y, MDD_ref$cumDensity, xlab = "DOY", ylab = "NMDI", font.lab = 2, breaks = c(0, 0.5, 0.75, 0.9, 0.95), col = grDevices::heat.colors(n = 4, alpha = 0.6))
contour(MDD_ref$x, MDD_ref$y, MDD_ref$cumDensity, levels = c(0, 0.5, 0.75, 0.9, 0.95), add = T, col = grDevices::grey(0.25), labcex = 1)
lines(seq(1, 365), MDD_ref$MAXY, lwd = 3, col = "dark red")

}

}
\seealso{
\code{\link{PLUGPhenAnoRFDMapPLUS}} \code{\link{PLUGPhenAnoRFDPLUS}}
}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
