% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avocado_WallToWallMap.R
\name{PlugPhenAnoProbMapPLUS}
\alias{PlugPhenAnoProbMapPLUS}
\title{Avocado algorithm - Wall-to-wall map version}
\usage{
PlugPhenAnoProbMapPLUS(
  s,
  phen,
  dates,
  h,
  anop,
  nCluster,
  outname,
  format,
  datatype,
  rge
)
}
\arguments{
\item{s}{Anomaly and probability brick created in section 2}

\item{phen}{Numeric vector with the values of the reference vegetation}

\item{dates}{The julian dates for each scene in the RasterBrick (x). Should be in the following format: "YYYY-MM-DD"}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of layers that are in x. For those layers the anomalies and probabilities will be calculated based on the phen. For example a raster stack has 450 layers, so anop=c(1:450).}

\item{nCluster}{Numeric. Number of CPU's to be used for the job.}

\item{outname}{Character vector with the output path and filename with extension or only the filename and extension if work directory was set. More information: See writeRaster}

\item{format}{Character. Output file type. More information: See writeFormats.}

\item{loc}{Location. SpatialPolygons overlapping one or multiple pixels.}

\item{x}{RasterBrick. Image time series brick}

\item{p}{The probability (%) of the anomalies being a candidate disturbance/change data point}

\item{cdates}{Sets the number of consecutive data points for a change to be detected (minimum 2 and maximum 5 data points).}

\item{rth}{sets a threshold (number of days) in which if a regrowth is detected within n days after a potential disturbance, then the candidate disturbance date is neglected.}

\item{dth}{sets a threshold (number of days) in which if a disturbance is detected within n days after a potential regrowth, then the candidate regrowth date is neglected.}
}
\description{
Continuous vegetation change detection
}
\details{
The output will be double the amount of layers as in x. The first layers are all the anomalies corresponding to the data points, the following layers are all the probabilities corresponding to the data points

The output will consist of 16 bands: Band1=disturbance detection; Band2=corresponding anomalies band 1; Band3=regrowth detection; Band4=corresponding anomalies band 3; Band5= 2nd disturbance detection for data points that had a previous disturbance and regrowth; Band 6 etc.
}
\section{1}{
 Create the reference vegetatation curve (phen).
}

\section{2}{
 Calculate the Anomaly and probability values (based on the difference between the pixel values and reference vegetation)
}

\section{3}{
 Automatic detection of the disturbance and regrowth events
}

\author{
Roberto O. Chavez, Mathieu Decuyper
}
