% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avocado_SinglePixel.R
\name{PLUGPhenAnoProbPLUS}
\alias{PLUGPhenAnoProbPLUS}
\title{Avocado algorithm - Single pixel version}
\usage{
PLUGPhenAnoProbPLUS(x, phen, dates, h, anop, rge)
}
\arguments{
\item{x}{RasterBrick. Image time series brick}

\item{phen}{Numeric vector with the values of the reference vegetation}

\item{dates}{The julian dates for each scene in the RasterBrick (x). Should be in the following format: "YYYY-MM-DD"}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of layers that are in x. For those layers the anomalies and probabilities will be calculated based on the phen. For example a raster stack has 450 layers, so anop=c(1:450).}

\item{loc}{Location. SpatialPolygons overlapping one or multiple pixels.}

\item{s}{Anomaly and probability time-series created in section 2}

\item{p}{The probability (in percentage) of the anomalies being a candidate disturbance/change data point}

\item{cdates}{Sets the number of consecutive data points for a change to be detected (minimum 2 and maximum 5 data points).}

\item{rth}{sets a threshold (number of days) in which if a regrowth is detected within n days after a potential disturbance, then the candidate disturbance date is neglected.}

\item{dth}{sets a threshold (number of days) in which if a disturbance is detected within n days after a potential regrowth, then the candidate regrowth date is neglected.}
}
\description{
Continuous vegetation change detection
}
\section{1}{
 Create the reference vegetatation curve (phen).
}

\section{2}{
 Calculate the Anomaly and probability values (based on the difference between the pixel values and reference vegetation)
}

\section{3}{
 Automatic detection of the disturbance and regrowth events
}

\examples{
# load MDD dataset
data(MDD)
data(RefForest)

############ Section 1: Reference vegetatation curve ################### 

x <- brick("MDD")
lan.info <- getSceneinfo(names(x))
lan.dates <-as.Date(lan.info$date)
loc <- readOGR(dsn=path.expand("YourDirectory"), layer="RefForest")
#loc <- spTransform(loc, "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0") #in case of different coordinate systems
ref.ext<-extent(loc)
ref.brick <- crop(x,ref.ext)
fin <- nrow(ref.brick)*ncol(ref.brick)
phen <- extract(ref.brick,1)
d1 <- lan.dates
for(i in 2:fin) {
 pp <-extract(ref.brick,i)
 phen <-c(phen,pp)
 d1 <-c(d1,lan.dates)
}
PhenKplot(phen,d1,h=1,nGS=365, xlab="DOY",ylab="NDMI",rge=c(0,10000))

Alternative: to take the median of all the pixel observations per date (Be aware: smoother curve but loss of natural variation)
ts.med <- NULL
for (i in 1:nlayers(ref.brick)) {
  if (all(is.na(getValues(ref.brick[[i]])))) {
    median <- NA
  } else {
    nn <- ncell(ref.brick[[i]])
    ss <- sampleRegular(ref.brick[[i]], size=nn)
    median <- median(ss, na.rm=T)
  }
  ts.med <- c(ts.med,median)
}
phen <- round(ts.med)
#' PhenKplot(phen,d1,h=1,nGS=365, xlab="DOY",ylab="NDMI",rge=c(0,10000))
############ Section 2: Calculate anomaly and probability values ###################
#Function - PLUGPhenAnoProbPLUS

}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
