% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Automatically create a directory and save a plot.}
\usage{
save_plot(p, f, w = 7, h = 7)
}
\arguments{
\item{p}{A plot.}

\item{f}{Save to this file.}

\item{w}{Width}

\item{h}{Height}
}
\value{
None
}
\description{
Automatically create a directory and save a plot.
}
\examples{

p1 <- ggpubr::ggboxplot(
data = ToothGrowth
, x = "supp"
, y = "dose"
, color = "supp"
)

save_plot(p = p1, f = "Figures/01.pdf", w = 5, h = 8)

}
