% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLUGPhenAnoRFDPLUS.R
\name{PLUGPhenAnoRFDMapPLUS}
\alias{PLUGPhenAnoRFDMapPLUS}
\title{Avocado algorithm - Wall-to-wall map version}
\usage{
PLUGPhenAnoRFDMapPLUS(s, phen, dates, h, anop, nCluster, outname, datatype,
  rge)
}
\arguments{
\item{s}{RasterStack time series of vegetation index (e.g. NDVI, EVI, NDMI)}

\item{phen}{Numeric vector with the values of the reference vegetation}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of values that are in x. For those values the anomalies and likelihoods will be calculated based on the phen. For example a time series has 450 values, so anop=c(1:450).}

\item{nCluster}{Numeric. Number of CPU's to be used for the job.}

\item{outname}{Character vector with the output path and filename with extension or only the filename and extension if work directory was set. More information: See writeRaster}

\item{datatype}{Character vector that determines the interpretation of values written to disk. More information: See \code{\link{writeFormats}}}

\item{rge}{A vector containing minimum and maximum values of the response variable used in the analysis. We suggest the use of theoretically based limits. For example in the case of MODIS NDVI or EVI, it ranges from 0 to 10,000, so rge =c(0,10000)}
}
\value{
RasterStack of nlayers([s]stack*2) containing all anomalies, followed by their likelihoods
}
\description{
Calculate the Anomaly and their likelihood values (based on the difference between the pixel values and reference vegetation) for a RasterStack
}
\examples{
\donttest{
source("PLUGPhenAnoRFDMapPLUS_20190902.R") # Load in the mapping function
dates <- lan.dates # The dates from your time-series brick (x)
PLUGPhenAnoRFDMapPLUS(s = MDD, dates = dates, h = 1, phen = phen, anop = c(1:n), nCluster = 1, outname = "YourDirectory/Filename.tif", format = "GTiff", datatype = "INT2S", rge = c(0, 10000))
}
}
\seealso{
\code{\link{PLUGPhenAnoRFDPLUS}}
}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
