% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLUGPhenAnoRFDPLUS.R
\name{PLUGPhenAnoRFDPLUS}
\alias{PLUGPhenAnoRFDPLUS}
\alias{PLUGPhenAnoRFDPlus}
\title{Avocado algorithm - Single pixel version}
\usage{
PLUGPhenAnoRFDPlus(x, phen, dates, h, anop, rge)
}
\arguments{
\item{x}{numeric vector. Time series of vegetation index (e.g. NDVI, EVI, NDMI)}

\item{phen}{Numeric vector with the values of the reference vegetation (all pixels of the reference area are considered)}

\item{dates}{A date vector. The number of dates must be equal to the number of values of time series.}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of values that are in x. For those values the anomalies and likelihoods will be calculated based on the phen. For example a time series has 450 values, so anop=c(1:450).}

\item{rge}{A vector containing minimum and maximum values of the response variable used in the analysis. We suggest the use of theoretically based limits. For example in the case of MODIS NDVI or EVI, it ranges from 0 to 10,000, so rge =c(0,10000)}
}
\value{
vector of length([x]*2) containing all anomalies, followed by their position within the reference frequency distribution (RFD)
}
\description{
Calculate the Anomaly and their likelihood values (based on the difference between the pixel values and reference vegetation) for a numeric vector
}
\examples{
\donttest{
#==========================================================================================================
# Reading landsat data and getting scene dates 
ndmibrick <- rast("~/Dropbox/GITHUB/avocado_testing_site/Pinus_test1_Landsat_NDMI_2000-01-01_2022-12-31.tif")#Load in your raster stack with the dates
#Load in the csv file obtained in GEE - beware column with the dates and "date" as header
dates.table <- read.csv('~/Dropbox/GITHUB/avocado_testing_site/Pinus_test1_Date_Landsat_NDMI_2000_2022.csv')
lan.dates <- as.Date(dates.table$date, format='\%d-\%m-\%Y') #change order to '\%Y-\%m-\%d'if you only get NA. Make sure that in this case there is a column name 'dates' in the csv file

#===================================================================================================================
#Create the reference curve
ref.core.shp <- vect("~/Dropbox/GITHUB/avocado_testing_site/Pinus_RefFor1.shp")
plot(ndmibrick[[1]])
plot(ref.core.shp,add=T)
ref.brick <- crop(ndmibrick,ref.core.shp)
ref.brick <- mask(ref.brick,ref.core.shp)
plot(ref.brick[[1]])
plot(ref.core.shp,add=T)
fin <- nrow(ref.brick)*ncol(ref.brick)
p2 <- as.numeric(extract(ref.brick,1))
d1 <- lan.dates
for(i in 2:fin) {
  pp <-as.numeric(extract(ref.brick,i))
  p2 <-c(p2,pp)
  d1 <-c(d1,lan.dates)
}

PhenKplot(x=p2,d1,h=1, xlab="DOY",ylab="NDMI",rge=c(0,8000))
#=========================================================================
# Task 2, calculating ano-prob + dist-regrowth FOR A SINGLE PIXEL (to test out the parameters before running large areas)
#Pixel location
plot(ndmibrick[[1]])
crs.data <- crs(ndmibrick)
xy <- rbind(c(4.918970,51.478009))#Drought detection in 2018
p <- vect(xy, crs=crs.data)
plot(p, add=T)
ts.inter <- as.numeric(extract(ndmibrick,p))
plot(ts.inter, ylim=c(-2000,8000))
#=========================================================================
}

}
\seealso{
\code{\link{PLUGPhenAnoRFDMapPLUS}}
}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
