% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLUGPhenAnoRFDPLUS.R
\name{PLUGPhenAnoRFDPLUS}
\alias{PLUGPhenAnoRFDPLUS}
\title{Avocado algorithm - Single pixel version}
\usage{
PLUGPhenAnoRFDPLUS(x, phen, dates, h, anop, rge)
}
\arguments{
\item{x}{numeric vector. Time series of vegetation index (e.g. NDVI, EVI, NDMI)}

\item{phen}{Numeric vector with the values of the reference vegetation}

\item{dates}{A date vector. The number of dates must be equal to the number of values of time series.}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of values that are in x. For those values the anomalies and likelihoods will be calculated based on the phen. For example a time series has 450 values, so anop=c(1:450).}

\item{rge}{A vector containing minimum and maximum values of the response variable used in the analysis. We suggest the use of theoretically based limits. For example in the case of MODIS NDVI or EVI, it ranges from 0 to 10,000, so rge =c(0,10000)}
}
\value{
vector of length([x]*2) containing all anomalies, followed by their likelihoods
}
\description{
Calculate the Anomaly and their likelihood values (based on the difference between the pixel values and reference vegetation) for a numeric vector
}
\examples{
\donttest{
#' # load MDD datasets
MDD <- stack(system.file("extdata", "MDD_NDMI_1990_2020.tif", package = "AVOCADO"))
load(system.file("extdata", "MDDref.RData", package = "AVOCADO"))
############ Section 1: Reference vegetatation curve ################### 
lan.info <- getSceneinfo(names(MDD))
lan.dates <-as.Date(lan.info$date)
MDDref <- readOGR(dsn=path.expand("YourDirectory"), layer="MDDref")
#MDDref <- spTransform(MDDref, "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0") #in case of different coordinate systems
ref.ext<-extent(MDDref)
ref.brick <- crop(MDD,ref.ext)
fin <- nrow(ref.brick)*ncol(ref.brick)
phen <- extract(ref.brick,1)
d1 <- lan.dates
for(i in 2:fin) {
 pp <-extract(ref.brick,i)
 phen <-c(phen,pp)
 d1 <-c(d1,lan.dates)
}
PhenKplot(phen,d1,h=1,nGS=365, xlab="DOY",ylab="NDMI",rge=c(0,10000))

Alternative: to take the median of all the pixel observations per date (Be aware: smoother curve but loss of natural variation)
ts.med <- NULL
for (i in 1:nlayers(ref.brick)) {
  if (all(is.na(getValues(ref.brick[[i]])))) {
    median <- NA
  } else {
    nn <- ncell(ref.brick[[i]])
    ss <- sampleRegular(ref.brick[[i]], size=nn)
    median <- median(ss, na.rm=T)
  }
  ts.med <- c(ts.med,median)
}
phen <- round(ts.med)
PhenKplot(phen,d1,h=1,nGS=365, xlab="DOY",ylab="NDMI",rge=c(0,10000))

############ Section 2:Calculate anomaly and their likelihoods values ###################
sample.rts <- rts(MDD,lan.dates)
pix.num <- cellFromXY(sample.rts,c(X,Ycoordinates))
ts.inter <- extract(sample.rts,pix.num)
vec <- as.vector(ts.inter)
dd <- lan.dates # NON SORTED DATES
nn <- seq(1,length(vec),1)
df <-data.frame(cbind(nn,dd))
sort <-df[order(as.Date(df$dd)),]
sort$vec <- vec
sort$date <- as.Date(sort$dd)
corr <- seq(1,length(vec),1)
sort$corr <- corr
x <- sort$vec
dates <- as.Date(sort$date)
ano_rfd <- PLUGPhenAnoRFDPLUS(x=x,phen=phen,dates=dates,h=1,anop=c(1:nlayers),rge=c(1,10000))
}
}
\seealso{
\code{\link{PLUGPhenAnoRFDMapPLUS}}
}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
