% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLUGPhenAnoRFDPLUS.R
\name{PLUGPhenAnoRFDPLUS}
\alias{PLUGPhenAnoRFDPLUS}
\alias{PLUGPhenAnoRFDPlus}
\title{Avocado algorithm - Single pixel version}
\usage{
PLUGPhenAnoRFDPlus(x, phen, dates, h, anop, rge)
}
\arguments{
\item{x}{numeric vector. Time series of vegetation index (e.g. NDVI, EVI, NDMI)}

\item{phen}{Numeric vector with the values of the reference vegetation (all pixels of the reference area are considered)}

\item{dates}{A date vector. The number of dates must be equal to the number of values of time series.}

\item{h}{Numeric. Geographic hemisphere to define the starting date of the growing season. h=1 Northern Hemisphere; h=2 Southern Hemisphere.}

\item{anop}{Numeric vector with the number of values that are in x. For those values the anomalies and likelihoods will be calculated based on the phen. For example a time series has 450 values, so anop=c(1:450).}

\item{rge}{A vector containing minimum and maximum values of the response variable used in the analysis. We suggest the use of theoretically based limits. For example in the case of MODIS NDVI or EVI, it ranges from 0 to 10,000, so rge =c(0,10000)}
}
\value{
vector of length([x]*2) containing all anomalies, followed by their position within the reference frequency distribution (RFD)
}
\description{
Calculate the Anomaly and their likelihood values (based on the difference between the pixel values and reference vegetation) for a numeric vector
}
\examples{
\dontrun{
# ===================================================================================================================
# Loading raster data
MDD <- rast(system.file("extdata", "MDD_NDMI_1990_2020.grd", package = "AVOCADO"))
# load dates vector
load(system.file("extdata", "MDD_dates.RData", package = "AVOCADO"))
# load  reference forest shapefile
load(system.file("extdata", "MDDref.RData", package = "AVOCADO"))
## time series extraction for a single pixel
px <- vect(cbind(-69.265, -12.48))
plot(MDD[[1]])
plot(px, add = T)

# extract series
px_series <- as.numeric(terra::extract(MDD, px, ID = F))
plot(MDD_dates, px_series, type = "l")
# Create the reference curve
ref.ext <- ext(MDDref)
ref.brick <- crop(MDD, ref.ext)
fin <- nrow(ref.brick) * ncol(ref.brick)
phen <- as.numeric(terra::extract(ref.brick, 1))
d1 <- MDD_dates
for (i in 2:fin) {
  pp <- as.numeric(terra::extract(ref.brick, i))
  phen <- c(phen, pp)
  d1 <- c(d1, MDD_dates)
}
PhenKplot(phen, d1, h = 1, xlab = "DOY", ylab = "NDMI", rge = c(0, 10000))
## Anomaly calculation
anom_rfd <- PLUGPhenAnoRFDPlus(x = px_series, phen = phen, dates = MDD_dates, h = 2, anop = c(1:1063), rge = c(1, 10000))
}

}
\seealso{
\code{\link{PLUGPhenAnoRFDMapPLUS}}
}
\author{
Roberto O. Chavez, Mathieu Decuyper
}
